+++ Runtime/Scripts/KtxLoadContext.cs
@@ -1,4 +1,4 @@
-﻿#if GLTFAST_BASISU
+﻿#if KTX_UNITY
 
 using System.Collections;
 using KtxUnity;
@@ -24,4 +24,4 @@ namespace GLTFast {
         }
     }
 }
-#endif // GLTFAST_BASISU
+#endif // KTX_UNITY