+++ Runtime/Scripts/KtxLoadNativeContext.cs
@@ -1,4 +1,4 @@
-﻿#if GLTFAST_BASISU
+﻿#if KTX_UNITY
 
 using System.Collections;
 using KtxUnity;
@@ -21,4 +21,4 @@ namespace GLTFast {
         }
     }
 }
-#endif // GLTFAST_BASISU
+#endif // KTX_UNITY