+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Support for Draco mesh compression is now optional (install DracoUnity package to enable it)
+- Support for KTX2/Basis Universal textures is now optional (install KtxUnity package to enable it)
 
 ## [0.11.0] - 2020-03-07
 ### Added