+++ ChangeLog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.0] - 2020-03-13
 ### Changed
 - Support for Draco mesh compression is now optional (install DracoUnity package to enable it)
 - Support for KTX2/Basis Universal textures is now optional (install KtxUnity package to enable it)